+++ sydtest/output-test/Spec.hs
@@ -204,6 +204,8 @@ spec = do
         context "Context2" $
           context "Context3" $
             True `shouldBe` False
+    it "shows a context when an exception is thrown as well" $
+      context "context" (undefined :: IO ())
   modifyMaxSize (`div` 10) $
     describe "Property" $ do
       describe "0 tests run" $ modifyMaxSuccess (const 0) $ it "shows a red '0 tests' when no tests are run" $ property $ \b -> b `shouldBe` False