+++ sydtest-hedgehog/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.0.0] - 2022-04-28
+
+### Added
+
+* Compatibility with `sydtest >=0.10`
+
 ## [0.1.0.0] - 2022-04-13
 
 ### Changed